cobol: Fix up cobol/{charmaps,valconv}.cc rules
sed -i is not portable, it is supported by GNU sed and perhaps some BSDs,
but not elsewhere.
Furthermore, I think it is far better to always use
 #include "../../libgcobol/something.h"
paths rather than something depending on the build directory.
And because we require GNU make, we don't have to have two different
rules for those, can use just one for both.
The l variable in there is just to make it fit into 80 columns.

2025-03-28  Jakub Jelinek  <jakub@redhat.com>

	* Make-lang.in (cobol/charmaps.cc, cobol/valconv.cc): Used sed -e
	instead of cp and multiple sed -i commands.  Always prefix libgcobol
	header names in #include directives with ../../libgcobol/ rather than
	something depending on $(LIB_SOURCE).