c++/modules: unexported friend template
Here we were failing to match the injected friend declaration to the
definition because the latter isn't exported.  But the friend is attached to
the module, so we need to look for any reachable declaration in that module,
not just the exports.

The duplicate_decls change is to avoid clobbering DECL_MODULE_IMPORT_P on
the imported definition; matching an injected friend doesn't change that
it's imported.  I considered checking get_originating_module == 0 or
!decl_defined_p instead of DECL_UNIQUE_FRIEND_P there, but I think this
situation is specific to friends.

I removed an assert because we have a test for the same condition a few
lines above.

gcc/cp/ChangeLog:

	* decl.cc (duplicate_decls): Don't clobber DECL_MODULE_IMPORT_P with
	an injected friend.
	* name-lookup.cc (check_module_override): Look at all reachable
	decls in decl's originating module.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/friend-9_a.C: New test.
	* g++.dg/modules/friend-9_b.C: New test.