testsuite: Fix up atomic-inst-ldlogic.c
r15-8956 changed in the test:
-/* { dg-final { scan-assembler-times "ldclr\t" 16} */
+/* { dg-final { scan-assembler-times "ldclr\t" 16 } */
which made it even worse than before, when the directive has
been silently ignored because it didn't match the regex for
directives.  Now it matches it but is unbalanced.

The following patch fixes it and adds space after all the
other scan-assembler-times counts in the file.

2025-03-30  Jakub Jelinek  <jakub@redhat.com>

	* gcc.target/aarch64/atomic-inst-ldlogic.c: Fix another
	unbalanced {} directive problem.  Add space after all
	scan-assembler-times counts.