c++: Honor noipa attribute for FE nothrow discovery [PR119518]
The following testcase has different code generation in bar depending on
whether foo is defined or just declared.
That is undesirable when it has noipa attribute, that attribute is
documented to be a black box between caller and callee, so the caller
shouldn't know about any implicitly determined properties of the callee
and callee shouldn't know about its callers.

E.g. the ipa-pure-const passes including nothrow discovery in there all
honor noipa attribute, but the FE did not.

2025-03-31  Jakub Jelinek  <jakub@redhat.com>

	PR c++/119518
	* decl.cc (finish_function): Don't set TREE_NOTHROW for
	functions with "noipa" attribute even when we can prove
	they can't throw.

	* g++.dg/opt/pr119518.C: New test.