arm: testsuite: fix vect-fmaxmin.c test
This is another case of a test that was both an executable test
requiring specific hardware and an assembler scan test.  The
requirement for the hardware was masking some useful testing that
could be done (by scanning the assembly output) on almost all test
runs.  Fixed in a similar manner to fmaxmin{,-2}.c by splitting the
test into two, one that scans the assembler output and one that
executes the compiled code if suitable hardware is available.

The masked issue was that this test was expecting vectorization to
occur that was incorrect given the options passed.  For correct
vectorization we need -funsafe-math-optimizations as the vector
version of the single-precision operation will apply a truncation of
denormal values.

gcc/testsuite/ChangeLog:

	* gcc.target/arm/vect-fmaxmin-2.c: New compile test.  Split from ...
	* gcc.target/arm/vect-fmaxmin.c: ... here.  Remove scan-assembler
	subtests.  For both, add -funsafe-math-optimizations.