c++: operator!= rewriting and arg-dep lookup
When considering an op== as a rewrite target, we need to disqualify it if
there is a matching op!= in the same scope.  But add_candidates was assuming
that we could use the same set of op!= for all op==, which is wrong if
arg-dep lookup finds op== in multiple namespaces.

This broke 20_util/optional/relops/constrained.cc if the order of the ADL
set changed.

gcc/cp/ChangeLog:

	* call.cc (add_candidates): Re-lookup ne_fns if we move into
	another namespace.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/spaceship-rewrite6.C: New test.