ada: VAST: Check basic tree properties
Check that the tree is really a tree, that parent pointers
make sense, that every node has been analyzed, and so on.
Most of these checks are disabled, because they fail in
many cases, including the compiler and run-time library.

Improve the debugging support in VAST. Walk subtrees "by hand",
rather than calling Atree.Traverse routines, because that
makes debugging printouts more convenient, and because we
want to keep a node stack for checking parents.

gcc/ada/ChangeLog:

	* vast.adb: Check basic tree properties.
	* atree.adb (Traverse_Field): Minor.
	* treepr.adb (Destroy): Minor comment.