libstdc++: Do not use compiler built-in in requires-clause
This is a partial backport of r14-6064-gc3f281a0c1ca50 which replaced a
use of __is_trivially_copyable in the requires-clause for std::bit_cast.
The change isn't actually necessary on the gcc-13 and gcc-12 branches,
but in gcc-14 using the built-in directly because ill-formed because we
started to mangle requires-clauses.

Replacing the built-in with the equivalent library trait makes it easier
to do bisections with preprocessed source created by GCC 13, because the
built-in won't cause errors with gcc-14 and later.

libstdc++-v3/ChangeLog:

	* include/std/bit (bit_cast): Use library trait instead of
	__is_trivially_copyable built-in.